requirements.txt
google-generativeai
googletrans==4.0.0-rc1
requests
Pillow
python-dotenv
